added group functionality

+++ app/src/main/java/fighting_mongooses/walkhealthy/ui/GroupActivity.java
@@ -3,11 +3,13 @@ package fighting_mongooses.walkhealthy.ui;
 import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.support.design.widget.FloatingActionButton;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
+import android.view.View;
 import android.widget.TableLayout;
 import android.widget.TableRow;
 import android.widget.TextView;
@@ -26,6 +28,13 @@ import fighting_mongooses.walkhealthy.objects.Group;
 import fighting_mongooses.walkhealthy.objects.User;
 import fighting_mongooses.walkhealthy.utilities.DatabaseTools;
 
+/**
+ * Group activity for group activities.
+ *
+ * This activity provides different actions for different groups.
+ *
+ * @author Mario Fellinger
+ */
 public class GroupActivity extends AppCompatActivity {
 
     public static final String KEY_EXTRA = "walkhealthy.KEY_GROUP";
@@ -40,7 +49,7 @@ public class GroupActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_group);
 
-        groupToolbar = (Toolbar) findViewById(R.id.groupToolbar);
+        groupToolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(groupToolbar);
 
         // add back arrow to toolbar
@@ -49,32 +58,46 @@ public class GroupActivity extends AppCompatActivity {
             getSupportActionBar().setDisplayShowHomeEnabled(true);
         }
 
+        FloatingActionButton myFab = (FloatingActionButton) findViewById(R.id.addGrpAcc);
+        myFab.setOnClickListener(new View.OnClickListener() {
+            public void onClick(View v) {
+                // CREATE A NEW GROUP EVENT
+            }
+        });
+
         memberlayout = (TableLayout) findViewById(R.id.memberlayout);
 
         if (getIntent().hasExtra(KEY_EXTRA)) {
-
-            DatabaseTools.readGroupData(getIntent().getStringExtra(KEY_EXTRA), new OnGetGroupListener() {
-                @Override
-                public void onStart() {
-                    // TODO BLOCK GUI WHILE GRP OBJECT IS LOADING
-                }
-
-                @Override
-                public void onSuccess(Group group) {
-                    GroupActivity.this.group = group;
-                    groupToolbar.setTitle(group.getName().toUpperCase());
-                    groupToolbar.setSubtitle("Walk Healthy Group");
-                    fetchUsers();
-                }
-
-                @Override
-                public void onFailed(DatabaseError databaseError) {
-
-                }
-            });
+            readGroupData();
         } else {
-            throw new IllegalArgumentException("Activity cannot find  extras " + KEY_EXTRA);
+            this.finish();
         }
+    }
+
+    /**
+     * Reads the group object from the database
+     * and fetches all group membe
+     */
+    private void readGroupData() {
+        DatabaseTools.readGroupData(getIntent().getStringExtra(KEY_EXTRA), new OnGetGroupListener() {
+            @Override
+            public void onStart() {
+                // TODO BLOCK GUI WHILE GRP OBJECT IS LOADING
+            }
+
+            @Override
+            public void onSuccess(Group group) {
+                GroupActivity.this.group = group;
+                groupToolbar.setTitle(group.getName().toUpperCase());
+                groupToolbar.setSubtitle("Walk Healthy Group");
+                fetchUsers();
+            }
+
+            @Override
+            public void onFailed(DatabaseError databaseError) {
+
+            }
+        });
     }
 
     /**
@@ -82,6 +105,7 @@ public class GroupActivity extends AppCompatActivity {
      * to the users lists.
      */
     private void fetchUsers() {
+        memberlayout.removeViews(0, memberlayout.getChildCount());
         for (Map.Entry<String, String> entry : group.getMembers().entrySet()) {
             DatabaseTools.readUserData(entry.getKey(), new OnGetUserListener() {
                 @Override
@@ -130,6 +154,19 @@ public class GroupActivity extends AppCompatActivity {
                 .setNegativeButton(android.R.string.no, null).show();
     }
 
+    /**
+     * Checks if the current user is group admin.
+     */
+    private boolean isCurrentUserAdmin() {
+        return DatabaseTools.getCurrentFirebaseUser().getUid().equals(group.getAdmin());
+    }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+        readGroupData();
+    }
+
     // Menu icons are inflated just as they were with actionbar
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
@@ -146,8 +183,22 @@ public class GroupActivity extends AppCompatActivity {
                 finish();
                 return super.onOptionsItemSelected(item);
 
+            case R.id.action_editgrp:
+                if (isCurrentUserAdmin()) {
+                    Intent intent = new Intent(GroupActivity.this, GroupEditActivity.class);
+                    intent.putExtra(GroupEditActivity.KEY_EXTRA, group.getName());
+                    startActivityForResult(intent, 0);
+                } else {
+                    Toast.makeText(GroupActivity.this, "Insufficient rights.", Toast.LENGTH_SHORT).show();
+                }
+                return true;
+
             case R.id.action_deletegrp:
-                deleteGroup();
+                if (isCurrentUserAdmin()) {
+                    deleteGroup();
+                } else {
+                    Toast.makeText(GroupActivity.this, "Insufficient rights.", Toast.LENGTH_SHORT).show();
+                }
                 return true;
 
             default:
